requirements.txt
Flask>=3.0
flask-caching>=2.3
httpx>=0.27
requests>=2.31
redis>=5.0
rq>=1.15
gunicorn>=21.2
gevent>=24.2
pytest>=8.0
ruff>=0.5
mypy>=1.10

beautifulsoup4>=4.12
Flask-Login>=0.6.3

# RAG / PDF support
PyMuPDF>=1.24
psycopg2-binary>=2.9
pgvector>=0.2
numpy>=1.24

# RAG Avancé
qdrant-client>=1.7.0
Pillow>=10.0.0

# LLM Providers SDKs
openai>=1.12
anthropic>=0.18
google-generativeai>=0.5
mistralai>=0.1
groq>=0.4
dashscope>=1.14
cryptography>=42.0
pytesseract==0.3.13
